+++ R/case_when.R
@@ -144,7 +144,7 @@ case_when <- function(...) {
   n <- length(fs)
 
   if (n == 0) {
-    abort("No cases provided")
+    abort("No cases provided.")
   }
 
   query <- vector("list", n)
@@ -188,7 +188,7 @@ validate_formula <- function(x, i, default_env, dots_env) {
     abort_case_when_formula(arg, i, x)
   }
   if (is_null(f_lhs(x))) {
-    abort("formulas must be two-sided")
+    abort("formulas must be two-sided.")
   }
 
   # Formula might be unevaluated, e.g. if it's been quosured
@@ -203,13 +203,13 @@ validate_formula <- function(x, i, default_env, dots_env) {
 abort_case_when_formula <- function(arg, i, obj) {
   deparsed <- fmt_obj1(deparse_trunc(arg))
   type <- friendly_type_of(obj)
-  abort(glue("Case {i} ({deparsed}) must be a two-sided formula, not {type}"))
+  abort(glue("Case {i} ({deparsed}) must be a two-sided formula, not {type}."))
 }
 
 abort_case_when_logical <- function(lhs, i, query) {
   deparsed <- fmt_obj1(deparse_trunc(quo_squash(lhs)))
   type <- friendly_type_of(query)
-  abort(glue("LHS of case {i} ({deparsed}) must be a logical vector, not {type}"))
+  abort(glue("LHS of case {i} ({deparsed}) must be a logical vector, not {type}."))
 }
 
 validate_case_when_length <- function(query, value, fs) {