+++ R/count-tally.R
@@ -83,7 +83,12 @@ count <- function(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = group_by
 tally <- function(x, wt = NULL, sort = FALSE, name = NULL) {
   n <- tally_n(x, {{ wt }})
   name <- check_name(x, name)
-  out <- summarise(x, !!name := !!n)
+
+  out <- local({
+    old.options <- options(dplyr.summarise.inform = FALSE)
+    on.exit(options(old.options))
+    summarise(x, !!name := !!n)
+  })
 
   if (sort) {
     arrange(out, desc(!!sym(name)))
@@ -144,16 +149,16 @@ tally_n <- function(x, wt) {
 check_name <- function(df, name) {
   if (is.null(name)) {
     if ("n" %in% tbl_vars(df)) {
-      glubort(
-        "Column 'n' is already present in output\n",
-        "* Use `name = \"new_name\"` to pick a new name"
-      )
+      abort(c(
+        "Column `n` is already present in output.",
+        i = "Use `name = \"new_name\"` to pick a new name."
+      ))
     }
     return("n")
   }
 
   if (!is.character(name) || length(name) != 1) {
-    abort("`name` must be a single string")
+    abort("`name` must be a single string.")
   }
 
   name