+++ R/lead-lag.R
@@ -55,7 +55,7 @@ lag <- function(x, n = 1L, default = NA, order_by = NULL, ...) {
 
   if (length(n) != 1 || !is.numeric(n) || n < 0) {
     bad_args("n", "must be a nonnegative integer scalar, ",
-      "not {friendly_type_of(n)} of length {length(n)}"
+      "not {friendly_type_of(n)} of length {length(n)}."
     )
   }
   if (n == 0) return(x)
@@ -80,7 +80,7 @@ lead <- function(x, n = 1L, default = NA, order_by = NULL, ...) {
 
   if (length(n) != 1 || !is.numeric(n) || n < 0) {
     bad_args("n", "must be a nonnegative integer scalar, ",
-             "not {friendly_type_of(n)} of length {length(n)}"
+             "not {friendly_type_of(n)} of length {length(n)}."
     )
   }
   if (n == 0) return(x)